# swagger/swagger.json
{
  "swagger": "2.0" ,
  "info": {
    "version": "1.0.0" ,
    "title": "Buddy App" ,
    "description": "Buddy Swagger" ,
    "license": {
      "name": "MIT" ,
      "url": "https://opensource.org/licenses/MIT"
    }
  } ,
  "host": "192.168.200.34:3000" ,
...
  "schemes": [
    "http"
  ] ,
  "consumes": [
    "application/json"
  ] ,
  "produces": [
    "application/json"
  ] ,
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security":[
    {
      "api_key":[]
    }
  ],
  "paths": {
...
    "/timeTable/addLecture":{
      "post":{
        "tags": [
          "TimeTable"
        ],
        "summary": "Add lecture",
        "description": "Allows admin to add lecturd",
        "parameters":[
          {
            "name": "lecture data",
            "in": "body",
            "description": "lecture body",
            "schema": {
              "$ref": "#/definitions/lectureData"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Lecture added",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/timeTable/{division_id}":{
      "get":{
        "tags":[
          "TimeTable"
        ],
        "summary":"Time table schedule",
        "description": "Get class wise time table schedule",
        "parameters": [
          {
            "name": "division_id",
            "in": "path",
            "required": true,
            "description": "ID of division",
            "type": "integer"
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Time table list",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/timeTable/{id}":{
      "delete": {
        "tags": [
          "TimeTable"
        ],
        "summary": "Delete lecture",
        "description" : "Delete specific lecture",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of time table",
            "type": "integer"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description": "Lecture deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put":{
        "tags": [
          "TimeTable"
        ],
        "summary": "Add lecture",
        "description": "Allows admin to add lecturd",
        "parameters":[
          {
            "name": "id",
            "in": "path",
            "description" :"Id of lecture to be edited",
            "required": true,
            "type": "integer"
          },
          {
            "name": "lecture data",
            "in": "body",
            "description": "lecture body",
            "schema": {
              "$ref": "#/definitions/lectureData"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Lecture updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/community/media": {
      "delete":{
        "tags": [
          "Community"
        ],
        "summary": "Delete media",
        "description": "Delete community media",
        "parameters":[
          {
            "name": "ids",
            "in": "body",
            "description": "ids of media to be removed",
            "schema":{
              "$ref": "#/definitions/mediaId"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Users removed changed",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/community/{id}":{
      "delete":{
        "tags":[
          "Community"
        ],
        "summary":"Community delete",
        "description": "Delete community related data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of community",
            "type": "integer"
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Community deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/studyMaterial/teacher/materials/{subject_id}/{division_id}":{
      "get":{
        "tags": [
          "Study Material"
        ],
        "summary": "Get materials",
        "description" : "Get materials posted by teacher",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "description": "ID of subject",
            "type": "integer"
          },
          {
            "name": "division_id",
            "in": "path",
            "required": true,
            "description": "ID of division",
            "type": "integer"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description": "Materials fetched",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/studyMaterial/media/{id}": {
      "delete": {
        "tags": [
          "Study Material"
        ],
        "summary": "Delete media",
        "description" : "Delete media uploaded by teacher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of media",
            "type": "integer"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description": "Materials deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/studyMaterial/student/materials/{subject_id}": {
      "get":{
        "tags": [
          "Study Material"
        ],
        "summary": "Get materials",
        "description" : "Get materials posted by teacher",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "description": "ID of subject",
            "type": "integer"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description": "Materials fetched",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  } ,
  "definitions": {
...
    "lectureData": {
      "required": [
        "day",
        "subject_id",
        "teacher_id",
        "duration_id",
        "standard_id",
        "division_id"
      ],
      "properties": {
        "day":{
          "type": "string"
        },
        "subject_id":{
          "type": "integer"
        },
        "teacher_id":{
          "type": "integer"
        },
        "duration_id":{
          "type": "integer"
        },
        "standard_id":{
          "type": "integer"
        },
        "division_id":{
          "type": "integer"
        }
      }
    },
...
    "mediaId" :{
      "required": [
        "ids"
      ],
      "properties" :{
        "ids" :{
          "type": "string"
        }
      }
    },
...
  }
}